Moved primitiveArray to its own module.
