+++ rust/arrow/src/array/mod.rs
@@ -86,6 +86,7 @@ mod array;
 mod array_binary;
 mod array_dictionary;
 mod array_list;
+mod array_primitive;
 mod array_string;
 mod array_struct;
 mod builder;
@@ -109,7 +110,6 @@ pub use self::data::ArrayData;
 pub use self::data::ArrayDataBuilder;
 pub use self::data::ArrayDataRef;
 
-pub use self::array::PrimitiveArray;
 pub use self::array_binary::BinaryArray;
 pub use self::array_binary::FixedSizeBinaryArray;
 pub use self::array_binary::LargeBinaryArray;
@@ -117,6 +117,7 @@ pub use self::array_dictionary::DictionaryArray;
 pub use self::array_list::FixedSizeListArray;
 pub use self::array_list::LargeListArray;
 pub use self::array_list::ListArray;
+pub use self::array_primitive::PrimitiveArray;
 pub use self::array_string::LargeStringArray;
 pub use self::array_string::StringArray;
 pub use self::array_struct::StructArray;
